Split up functions into several modules. Add PTW Analyzer. Add Frequency
PTW Analyzer compares srts found in a folder with the frequency yomichan
dictionary and calculate a value of how valuable the vocabulary inside
the srts is. It also compares it against just the vocab in the vocab
file.

Added the capability to include frequency data of a yomichan frequency
dictionary. See my other project: yomidict